ML/ml/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
groq==0.25.0
python-dotenv==1.0.1
